% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.MCMC.SMCM.R
\name{fit.MCMC.SMCM}
\alias{fit.MCMC.SMCM}
\title{Fit Bayesian Semi-parametric Mixture Cure Model (SMCM) via MCMC algorithm}
\usage{
fit.MCMC.SMCM(
  data,
  hyperpar = list(),
  nchains = 2,
  nIter = 6000,
  warmup = 1000,
  thin = 20,
  mcmc.parallel = "parLapply",
  standardize,
  probs,
  save_loglik,
  seed = NULL
)
}
\arguments{
\item{data}{a list containing observed data from \code{n} subjects,
\code{observed_time}, \code{delta}, \code{X}, \code{Z}. See details for more information.}

\item{hyperpar}{a list containing prior parameter values for the SMCM model; among c("r1", "delta1", "r2", "delta2", "a", "b").
See details for more information.}

\item{nchains}{Number of MCMC chains.}

\item{nIter}{Number of MCMC iterations.}

\item{warmup}{Number of warmup iterations.}

\item{thin}{Thinning interval for MCMC samples.}

\item{mcmc.parallel}{Parallelization method to run nchains for MCMC algorithm, can be \code{lapply},
\code{foreach} from \pkg{doParallel} package or \code{parLapply} from \pkg{parallel} package.}

\item{standardize}{Logical indicating whether to standardize covariates.}

\item{probs}{A numeric vector of quantiles used to define the cut points of the piecewise exponential baseline hazard function.
For example, use \code{probs = 0} when \eqn{J = 1}, \code{probs = 0.5} or any other quantile value when \eqn{J = 2},
and a vector such as \code{probs = c(0.35, 0.70)} when \eqn{J = 3}.}

\item{save_loglik}{Integer indicating whether to save log-likelihood values (0 = FALSE, 1 = TRUE).}

\item{seed}{Random seed for reproducibility.}
}
\value{
A list containing:
\item{fit}{A list containing all results across MCMC chains.}
\item{fit.results}{Summary of posterior estimation results, including posterior means, credible intervals, and
convergence diagnostics for all model parameters.}
\item{b.chains}{Posterior samples for the regression coefficients \eqn{\bm b} across all MCMC chains.}
\item{beta.chains}{Posterior samples for the regression coefficients \eqn{\bm \beta} across all MCMC chains.}
\item{lambda.chains}{Posterior samples for the regression coefficients \eqn{\bm \lambda} across all MCMC chains.}
\item{Rhat}{Gelman–Rubin convergence statistic for assessing MCMC chain convergence.}
\item{loo.results}{Model fit statistics including DIC and LPML when \code{save_loglik} = TRUE.}
\item{loglik.matrix.chains}{Matrix of the log-likelihood of the model over MCMC samples when \code{save_loglik = TRUE}.}
\item{data}{Processed survival data used for model fitting.}
\item{priors}{Prior parameter settings used in the model.}
\item{mcmc.info}{nchains, nIter, warmup and thin are used in the model.}
}
\description{
This is the main function to fit the Bayesian SMCM model using MCMC algorithm in Kızılaslan and Vitelli (2025).
}
\details{
fit.MCMC.SMCM

\tabular{ll}{
\code{observed_time} \tab a vector of \eqn{n} times to the event.\cr
\code{delta} \tab a vector of \eqn{n} censoring indicators for the event time (1=event occurred, 0=censored).\cr
\code{X} \tab covariate matrix for the incidence part of the model \eqn{\bm X}, \eqn{n} observations by \eqn{p_2} variables,
where \eqn{p_2} is the length of \eqn{\bm \beta}.\cr
\code{Z} \tab covariate matrix for the latency part of the model \eqn{\bm Z}, \eqn{n} observations by \eqn{p_1+1} variables,
where \eqn{p_1+1} is the length of \eqn{\bm \b}.\cr
\code{r1} \tab the shape parameter of the gamma prior for \eqn{\eta^2}.\cr
\code{delta1} \tab the rate parameter of the gamma prior for \eqn{\eta^2}.\cr
\code{r2} \tab the shape parameter of the gamma prior for \eqn{\eta^{*2}}.\cr
\code{delta2} \tab the rate parameter of the gamma prior for \eqn{\eta^{*2}}.\cr
\code{a} \tab the shape parameter of the gamma prior for \eqn{\bm \lambda}.\cr
\code{b} \tab the rate parameter of the gamma prior for \eqn{\bm \lambda}.\cr
}
}
\examples{
# Fit SMCM model to data generated from Scenario 1 in our simulation study
n <- 300
b.true <- c( 0.4, 0.5, 0.1)
beta.true <- c(1, 0.2 )
baseline.hazard.rates <- 1
intervals <- c(0, 16)
prob.cov.X <- c(0.5)
prob.cov.Z <- c(0.5)
dat1 <- simSMCM(n, b.true, beta.true, baseline.hazard.rates, intervals, seed = 2025, 
  cens.start = 30/365, cens.end = 32,  prob.cov.X, prob.cov.Z, same.cov = TRUE )
str(dat1)
nchains = 2; nIter = 2500; warmup = 500; thin = 10
priorPar = list( r1 = 1, delta1 = 0.0001, r2 = 1, delta2 = 0.0001, a = 0.1, b = 0.1 )
out.smcm.mcmc = fit.MCMC.SMCM ( data = dat1, hyperpar = priorPar, nchains, nIter, warmup, 
  thin, mcmc.parallel = "parLapply", standardize = FALSE, probs = 0, save_loglik = 1, seed = 2025 )
print_smcm( out.smcm.mcmc, stan.model = FALSE, frailty = FALSE )

}
\references{
Kızılaslan, F., Vitelli, V. (2025). Bayesian Semiparametric Mixture Cure (Frailty) Model, arXiv \url{https://arxiv.org/???}.
}
