% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.interval.mcmc.R
\name{HPD.interval.mcmc}
\alias{HPD.interval.mcmc}
\title{Highest Posterior Density (HPD) intervals for all MCMC chains of SMCM and SMCFM.}
\usage{
HPD.interval.mcmc(fit.mcmc.chains, pars, warmup, thin, ...)
}
\arguments{
\item{fit.mcmc.chains}{MCMC results for all the chains based on our MCMC results.}

\item{pars}{Parameters can be "beta.p" or "b.p" or "lambda.p".}

\item{warmup}{Number of warmup iterations.}

\item{thin}{Thinning interval for MCMC samples.}

\item{...}{If you have true values of the parameters, give it as "pars.true", then cp value return as 0 and 1
for each parameter with respect to HPS interval includes true value or not.}
}
\value{
A data frame containing the lower and upper bounds of the 95\% HPD credible interval for each parameter.
If true parameter values are provided, an additional column 'cp' indicates whether the true value falls within
the interval (1) or not (0).
}
\description{
It is calculated by using \pkg{coda} via \link[coda]{HPDinterval}.
}
\examples{
\dontrun{
# Assuming fit.MCMC.SMCM is fitted as "out", then
HPD.interval.mcmc(fit.mcmc.chains = out$fit, pars = "b.p", warmup = 1, thin = 1)
}
}
