% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_SMCFM.R
\name{logLik_SMCFM}
\alias{logLik_SMCFM}
\title{Logarithm of the likelihood for SMCFM}
\usage{
logLik_SMCFM(survObj, priorPara, b.coef, beta.coef, lambda.coef, theta.coef)
}
\arguments{
\item{survObj}{An object of class SMCFM containing survival data and covariates.}

\item{priorPara}{A list of prior parameters including interval bounds.}

\item{b.coef}{regression coefficients for the incidence part of the model.}

\item{beta.coef}{regression coefficients for the latency part of the model.}

\item{lambda.coef}{baseline hazard parameters for the piecewise exponential distribution.}

\item{theta.coef}{gamma distribution parameter for the frailty distribution.}
}
\value{
Numeric value of the log-likelihood.
}
\description{
Compute the logarithm of the likelihood of SMCFM for the right-censored observed survival data.
}
