% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_SMCM.R
\name{logLik_SMCM}
\alias{logLik_SMCM}
\title{Logarithm of the likelihood for SMCM}
\usage{
logLik_SMCM(survObj, priorPara, b.coef, beta.coef, lambda.coef)
}
\arguments{
\item{survObj}{An object of class SMCM containing survival data and covariates.}

\item{priorPara}{A list of prior parameters including interval bounds.}

\item{b.coef}{regression coefficients for the incidence part of the model.}

\item{beta.coef}{regression coefficients for the latency part of the model.}

\item{lambda.coef}{baseline hazard parameters for the piecewise exponential distribution.}
}
\value{
Numeric value of the log-likelihood.
}
\description{
Compute the logarithm of the likelihood of SMCM for the right-censored observed survival data.
}
