% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.SMCM.R
\name{MCMC.SMCM}
\alias{MCMC.SMCM}
\title{Function to MCMC algorithm for Semi-parametric Mixture Cure Model (SMCM)}
\usage{
MCMC.SMCM(survObj, priorPara, initial, nIter = 6000, save_loglik = 0)
}
\arguments{
\item{survObj}{a list containing observed data from \code{n} subjects,
\code{t}, \code{di}, \code{X}, \code{Z}. See details for more information.}

\item{priorPara}{a list containing prior parameter values for the SMCM model; among c("r1", "delta1", "r2", "delta2", "a", "b", "s", "J").
See details for more information.}

\item{initial}{Initial value of parameters.}

\item{nIter}{Number of MCMC iterations.}

\item{save_loglik}{Integer indicating whether to save log-likelihood values (0 = FALSE, 1 = TRUE).}
}
\value{
A list containing:
\item{beta.p}{MCMC samples for the regression coefficients \eqn{\bm \beta}.}
\item{b.p}{MCMC samples for the regression coefficients \eqn{\bm b}.}
\item{tauSq.p}{posterior samples for \eqn{\tau^2}.}
\item{tauStarSq.p}{posterior samples for \eqn{\tau^{*2}}.}
\item{lambda.p}{posterior samples for \eqn{\bm \lambda}.}
\item{mcmcOutcome}{The list containing posterior samples for the remaining model parameters.}
\item{t}{Processed times to the event data.}
\item{di}{Processed censoring indicators for the event time.}
\item{loglik.matrix}{Matrix of the log-likelihood of the model over MCMC samples when \code{save_loglik = TRUE}.}
}
\description{
This function obtains posterior samples for all model parameters using an MCMC algorithm
for a single chain when applying the \code{MCMC.SMCM} framework in K覺z覺laslan and Vitelli (2025) to the data.
}
\details{
The input data and prior parameters include the following:
\tabular{ll}{
\code{t}\tab A vector of \eqn{n} times to the event.\cr
\code{di}\tab A vector of \eqn{n} censoring indicators (1 = event occurred, 0 = censored).\cr
\code{X} \tab covariate matrix for the incidence part of the model \eqn{\bm X}, \eqn{n} observations by \eqn{p_2} variables,
where \eqn{p_2} is the length of \eqn{\bm \beta}.\cr
\code{Z} \tab covariate matrix for the latency part of the model \eqn{\bm Z}, \eqn{n} observations by \eqn{p_1+1} variables,
where \eqn{p_1+1} is the length of \eqn{\bm \b}.\cr
\code{r1}\tab Shape parameter of the gamma prior for \eqn{\eta^2}.\cr
\code{delta1}\tab Rate parameter of the gamma prior for \eqn{\eta^2}.\cr
\code{r2}\tab Shape parameter of the gamma prior for \eqn{\eta^{*2}}.\cr
\code{delta2}\tab Rate parameter of the gamma prior for \eqn{\eta^{*2}}.\cr
\code{a}\tab Shape parameter of the gamma prior for \eqn{\bm \lambda}.\cr
\code{b}\tab Rate parameter of the gamma prior for \eqn{\bm \lambda}.\cr
\code{s}\tab Interval boundaries for the piecewise exponential baseline hazard.\cr
\code{J}\tab Total number of intervals for the piecewise exponential baseline hazard.\cr
}
}
\references{
K覺z覺laslan, F., Vitelli, V. (2025). Bayesian Semiparametric Mixture Cure (Frailty) Model, arXiv \url{https://arxiv.org/???}.
}
