% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_trace.smcm.R
\name{mcmc_trace.smcm}
\alias{mcmc_trace.smcm}
\title{Trace plot of MCMC draws for SMCM and SMCFM.}
\usage{
mcmc_trace.smcm(fit.mcmc.chains, pars, warmup = 1, thin = 1)
}
\arguments{
\item{fit.mcmc.chains}{A list of MCMC results for all the chains based on our MCMC algorithms.}

\item{pars}{Parameters can be "beta.p" or "b.p" or "lambda.p" or "theta.p"}

\item{warmup}{Number of warmup iterations.}

\item{thin}{Thinning interval for MCMC samples.}
}
\value{
A trace plot of the specified parameters.
}
\description{
Trace plot of posterior MCMC draws of the parameters including \eqn{\bm \beta}, \eqn{\bm b}, and
\eqn{\bm \lambda} for SMCM and SMCFM based on our MCMC algorithms.
It is similar to \pkg{bayesplot} \link[bayesplot]{mcmc_trace} function but customized for our MCMC outputs.
}
\examples{
\dontrun{
# Assuming fit.MCMC.SMCM is fitted as "out", then
mcmc_trace.smcm(fit.mcmc.chains = out$fit, pars="b.p", warmup = 1, thin = 1 )
} 
}
