% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_smcm.R
\name{print_smcm}
\alias{print_smcm}
\title{Print the summary results from SMCM or SMCFM fitting based on MCMC or Stan}
\usage{
print_smcm(out, stan.model, frailty)
}
\arguments{
\item{out}{A list variable, an output from fit.MCMC.SMCM or fit.MCMC.SMCFM or any proposed Stan model.}

\item{stan.model}{Logical variable, if TRUE, it is from Stan, if FALSE, it is from our MCMC.}

\item{frailty}{Logical variable, if TRUE, it is cure frailty model, if FALSE, it is cure model.}
}
\value{
A data frame containing the summary results, including posterior mean, posterior SD,
95\% HPD credible interval, and Rhat for all model parameters.
}
\description{
This function prints the summary results of the fitted SMCM or SMCFM model based on MCMC or Stan.
It includes the posterior mean, posterior standard deviation, 95\% HPD credible interval, and Rhat for each parameter.
The function handles both MCMC outputs from custom implementations and outputs from Stan models.
It also calculates the posterior standard deviation and HPD interval for the cure probability \eqn{\pi(z)}.
}
\examples{
\dontrun{
# Assuming fit.MCMC.SMCM is fitted as "out", then
print_smcm(out = out.smcm.mcmc, stan.model = FALSE, frailty = FALSE)
}

}
