% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhat.mcmc.R
\name{Rhat.mcmc}
\alias{Rhat.mcmc}
\title{Gelman and Rubin's convergence diagnostic for MCMC chains of SMCM and SMCFM.}
\usage{
Rhat.mcmc(fit, warmup = 1, thin = 1)
}
\arguments{
\item{fit}{A list variable \code{fit} from the fitted model of MCMC.SMCM or MCMC.SMCFM output which includes posterior samples of
all the parameters over chains.}

\item{warmup}{Number of warmup iterations.}

\item{thin}{Thinning interval for MCMC samples.}
}
\value{
Rhat values for all model parameters \eqn{\bm b}, \eqn{\bm \beta}, \eqn{\bm \lambda} and/or \eqn{\theta}.
}
\description{
The ‘potential scale reduction factor’ is calculated by using \pkg{coda} via \link[coda]{gelman.diag}.
}
\examples{
\dontrun{
# Assuming fit.MCMC.SMCM or fit.MCMC.SMCFM is fitted as "out", then  
Rhat_values <- Rhat.mcmc(fit = out$fit, warmup, thin)
print(Rhat_values)
} 

}
