% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.posterior.smcm.R
\name{SD.posterior.smcm}
\alias{SD.posterior.smcm}
\title{Calculate posterior standard deviation of the estimates from SMCM or SMCFM fitting based on MCMC or Stan}
\usage{
SD.posterior.smcm(out, stan.model = FALSE, frailty = FALSE)
}
\arguments{
\item{out}{A list variable, an output from fit.MCMC.SMCM or fit.MCMC.SMCFM or any proposed Stan model.}

\item{stan.model}{Logical variable, if TRUE, it is from Stan, if FALSE, it is from our MCMC.}

\item{frailty}{Logical variable, if TRUE, it is cure frailty model, if FALSE, it is cure model.}
}
\value{
A numeric vector containing the posterior standard deviations for each parameter.
}
\description{
This function calculates the posterior standard deviation of the parameter estimates from the fitted SMCM
or SMCFM model based on MCMC or Stan.
It processes the MCMC chains or Stan outputs to compute the standard deviation for each parameter.
}
\examples{
\dontrun{
# Assuming fit.MCMC.SMCM is fitted as "out",  then
SD.posterior.smcm(out = out, stan.model = FALSE, frailty = FALSE)
}

}
