% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSMCM.R
\name{simSMCM}
\alias{simSMCM}
\title{Function to generate data from Semi-parametric Mixture Cure Model (SMCM)}
\usage{
simSMCM(
  n,
  b.true,
  beta.true,
  baseline.hazard.rates,
  intervals,
  seed,
  cens.start,
  cens.end,
  prob.cov.X,
  prob.cov.Z,
  same.cov
)
}
\arguments{
\item{n}{Sample size}

\item{b.true}{True regression coefficients for the incidence part of the model.}

\item{beta.true}{True regression coefficients for the latency part of the model.}

\item{baseline.hazard.rates}{Baseline hazard parameters for the piecewise exponential distribution.}

\item{intervals}{Vector of interval boundaries.}

\item{seed}{Random seed for reproducibility.}

\item{cens.start}{Start time for censoring distribution.}

\item{cens.end}{End time for censoring distribution.}

\item{prob.cov.X}{Vector of probabilities for generating binary covariates in \eqn{\bm X}, size at most length(beta.true).}

\item{prob.cov.Z}{Vector of probabilities for generating binary covariates in \eqn{\bm Z}, size at most length(b.true) - 1.}

\item{same.cov}{Logical indicating whether to use the same covariates for both \eqn{\bm X} and \eqn{\bm Z}.}
}
\value{
A list containing:
\item{b.true}{True regression coefficients for the incidence part of the model.}
\item{beta.true}{True regression coefficients for the latency part of the model.}
\item{baseline.hazard.rates}{Baseline hazard parameters for the piecewise exponential distribution.}
\item{X}{Covariate matrix for the incidence part of the model.}
\item{Z}{Covariate matrix for the latency part of the model.}
\item{observed_time}{Observed times to the event.}
\item{delta}{Censoring indicators for the event time (1=event occurred, 0=censored).}
\item{censor.rate}{Censoring rate for the simulated data.}
\item{intervals}{Vector of interval boundaries.}
}
\description{
Simulate data from Semi-parametric Mixture Cure Model (SMCM) with the piecewise exponential baseline hazard function.
}
\examples{

#1. Generate data as in Yin and Ibrahim (2005), i.e. Scenario 1 in the simulation study of
# Kızılaslan and Vitelli (2025).
n <- 300
b.true <- c( 0.4, 0.5, 0.1)
beta.true <- c(1, 0.2 )
baseline.hazard.rates <- 1
intervals <- c(0, 16)
prob.cov.X <- c(0.5)    
prob.cov.Z <- c(0.5)    
dat1 <- simSMCM(n, b.true, beta.true, baseline.hazard.rates, intervals, seed = 2026, 
  cens.start = 30/365, cens.end = 32,  prob.cov.X, prob.cov.Z, same.cov = TRUE )
str(dat1)

#2. Generate data as Scenario 2 in the simulation study of Kızılaslan and Vitelli (2025).
n <- 200
b.true <-  c( 0.25, -1, 1.5, 0.5)
beta.true <- c(-1, 0.5, 2)
baseline.hazard.rates <- c( 0.2, 0.15, 0.30)
intervals <- c( 0, 10, 20, 30) 
prob.cov.X <- c(0.5)    
prob.cov.Z <- c(0.6)    
dat2 <- simSMCM(n, b.true, beta.true, baseline.hazard.rates, intervals, seed = 2026, 
  cens.start = 30/365, cens.end = 32,  prob.cov.X, prob.cov.Z, same.cov = FALSE )
str(dat2)

#3. Generate data as Scenario 3 in the simulation study of Kızılaslan and Vitelli (2025).
n <- 200
b.true <- c( -0.5, 1, 1.5, -2)
beta.true <- c( 1.5, -0.30, 0.7, 1)
baseline.hazard.rates <- c( 0.15, 0.30, 0.50, 1)
intervals <- c( 0, 15, 30, 45, 60) 
prob.cov.X <- c(0.5, 0.25, 0.65)    
prob.cov.Z <- c(0.3, 0.6)   
dat3 <- simSMCM(n, b.true, beta.true, baseline.hazard.rates, intervals, seed = 2026, 
  cens.start = 2*(30/365), cens.end = 60,  prob.cov.X, prob.cov.Z, same.cov = FALSE )
str(dat3)

}
\references{
Yin, G., Ibrahim, J. G. (2005). Cure rate models: a unified approach. Canadian Journal of Statistics, 33(4):559–570.

Kızılaslan, F., Vitelli, V. (2025). Bayesian Semiparametric Mixture Cure (Frailty) Model, arXiv \url{https://arxiv.org/???}.
}
