% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_hazard.R
\name{cumulative_hazard}
\alias{cumulative_hazard}
\title{Cumulative hazard and hazard functions under the piecewise exponential baseline hazard model}
\usage{
cumulative_hazard(t, interval_bounds, lambda)
}
\arguments{
\item{t}{A vector of times to event.}

\item{interval_bounds}{Interval boundaries for the hazard distribution of piecewise exponential.}

\item{lambda}{Baseline hazard parameters of the piecewise exponential distribution.}
}
\value{
A list with \code{h0} (hazard function values) and \code{H0} (cumulative hazard function values).
}
\description{
Calculates the cumulative hazard and hazard functions defined by the piecewise exponential baseline hazard model.
}
